chore: clean engine codes
