--- apps/engine/src/config/database.ts
@@ -1,5 +1,0 @@
-import { parseDatabaseOptionsFromEnv } from '@tachybase/database';
-
-export async function parseDatabaseOptions() {
-  return await parseDatabaseOptionsFromEnv();
-}
