--- apps/engine/src/config/logger.ts
@@ -1,13 +1,0 @@
-import { getLoggerLevel, getLoggerTransport } from '@tachybase/logger';
-import { AppLoggerOptions } from '@tachybase/server';
-
-export default {
-  request: {
-    transports: getLoggerTransport(),
-    level: getLoggerLevel(),
-  },
-  system: {
-    transports: getLoggerTransport(),
-    level: getLoggerLevel(),
-  },
-} as AppLoggerOptions;
