--- apps/engine/src/config/plugins.ts
@@ -1,3 +1,0 @@
-import { PluginConfiguration } from '@tachybase/server';
-
-export default ['tachybase'] as PluginConfiguration[];
