--- apps/engine/src/config/resourcer.ts
@@ -1,3 +1,0 @@
-export default {
-  prefix: process.env.API_BASE_PATH,
-};
